+++ IPython/kernel/client.py
@@ -348,7 +348,8 @@ class KernelClient(ConnectionFileMixin):
     def _handle_kernel_info_reply(self, msg):
         """handle kernel info reply
 
-        sets protocol adaptation version
+        sets protocol adaptation version. This might
+        be run from a separate thread.
         """
         adapt_version = int(msg['content']['protocol_version'].split('.')[0])
         if adapt_version != major_protocol_version:
